updates
tried gpu acceleration - my machine's vram is too small for any meaningful tests. i'm defaulting to cpu and ram for now.

generalized the structure of the spatial kernel - we can have an arbitrary amount of encoder/decoder layers now. the memory complexity of this is around O(a^n) with a the pooling size especially during gradient application or for particularly expensive optimizers (AdamW...). this is exponential in the amount of layers and is therefore EXTREMELY memory-intensive for especially large or fine lattices, if we want good optimization.